+++ src/persist/src/nemesis/direct.rs
@@ -560,11 +560,14 @@ impl TimelyWorkers {
 mod tests {
     use mz_ore::metrics::MetricsRegistry;
     use tempfile::TempDir;
+    use tokio::runtime::Runtime as AsyncRuntime;
 
+    use crate::error::ErrorLog;
     use crate::file::{FileBlob, FileLog};
     use crate::mem::MemRegistry;
     use crate::nemesis::generator::GeneratorConfig;
     use crate::runtime::RuntimeConfig;
+    use crate::s3::{S3Blob, S3BlobConfig};
     use crate::storage::Blob;
     use crate::unreliable::{UnreliableBlob, UnreliableLog};
     use crate::{nemesis, runtime};
@@ -611,6 +614,37 @@ mod tests {
         // second, so run this one for fewer steps than the other tests. Revisit
         // once we pipeline write calls in Log.
         nemesis::run(10, GeneratorConfig::default(), direct);
+    }
+
+    impl StartRuntime for (Arc<AsyncRuntime>, S3BlobConfig) {
+        fn start_runtime(&mut self, unreliable: UnreliableHandle) -> Result<RuntimeClient, Error> {
+            let (runtime, config) = self;
+            let (runtime, config) = (Arc::clone(runtime), config.clone());
+            let guard = runtime.enter();
+            let blob = S3Blob::open_exclusive(config, ("reentrance0", "direct_s3").into())?;
+            drop(guard);
+            let blob = UnreliableBlob::from_handle(blob, unreliable);
+            runtime::start(
+                RuntimeConfig::for_tests(),
+                ErrorLog,
+                blob,
+                mz_build_info::DUMMY_BUILD_INFO,
+                &MetricsRegistry::new(),
+                Some(runtime),
+            )
+        }
+    }
+
+    #[test]
+    fn direct_s3() {
+        let runtime = Arc::new(AsyncRuntime::new().expect("failed to create async runtime"));
+        let config = runtime
+            .block_on(S3BlobConfig::new_for_test())
+            .expect("loading s3 config failed");
+        if let Some(config) = config {
+            let direct = Direct::new((runtime, config)).expect("initial start failed");
+            nemesis::run(10, GeneratorConfig::default(), direct);
+        }
     }
 
     // A variant with a traffic pattern vaguely like production usage of