+++ src/sql-parser/src/ast/defs/expr.rs
@@ -81,6 +81,14 @@ pub enum Expr<T: AstInfo> {
         subquery: Box<Query<T>>,
         negated: bool,
     },
+    /// `<expr> [ NOT ] {LIKE, ILIKE} <pattern> [ ESCAPE <escape> ]`
+    Like {
+        expr: Box<Expr<T>>,
+        pattern: Box<Expr<T>>,
+        escape: Option<Box<Expr<T>>>,
+        case_insensitive: bool,
+        negated: bool,
+    },
     /// `<expr> [ NOT ] BETWEEN <low> AND <high>`
     Between {
         expr: Box<Expr<T>>,
@@ -178,13 +186,8 @@ pub enum Expr<T: AstInfo> {
     /// `LIST[<expr>*]`
     List(Vec<Expr<T>>),
     ListSubquery(Box<Query<T>>),
-    /// `<expr>[<expr>]`
-    SubscriptScalar {
-        expr: Box<Expr<T>>,
-        subscript: Box<Expr<T>>,
-    },
-    /// `<expr>[<expr>:<expr>(, <expr>?:<expr>?)*]`
-    SubscriptSlice {
+    /// `<expr>([<expr>(:<expr>)?])+`
+    Subscript {
         expr: Box<Expr<T>>,
         positions: Vec<SubscriptPosition<T>>,
     },
@@ -261,6 +264,31 @@ impl<T: AstInfo> AstDisplay for Expr<T> {
                 f.write_str("IN (");
                 f.write_node(&subquery);
                 f.write_str(")");
+            }
+            Expr::Like {
+                expr,
+                pattern,
+                escape,
+                case_insensitive,
+                negated,
+            } => {
+                f.write_node(&expr);
+                f.write_str(match (*case_insensitive, *negated) {
+                    (false, false) => " ~~ ",
+                    (false, true) => " !~~ ",
+                    (true, false) => " ~~* ",
+                    (true, true) => " !~~* ",
+                });
+                match escape {
+                    Some(escape) => {
+                        f.write_str("like_escape(");
+                        f.write_node(&pattern);
+                        f.write_str(", ");
+                        f.write_node(escape);
+                        f.write_str(")");
+                    }
+                    None => f.write_node(&pattern),
+                }
             }
             Expr::Between {
                 expr,
@@ -452,16 +480,21 @@ impl<T: AstInfo> AstDisplay for Expr<T> {
                 f.write_node(&s);
                 f.write_str(")");
             }
-            Expr::SubscriptScalar { expr, subscript } => {
+            Expr::Subscript { expr, positions } => {
                 f.write_node(&expr);
                 f.write_str("[");
-                f.write_node(subscript);
-                f.write_str("]");
-            }
-            Expr::SubscriptSlice { expr, positions } => {
-                f.write_node(&expr);
-                f.write_str("[");
-                f.write_node(&display::comma_separated(positions));
+
+                let mut first = true;
+
+                for p in positions {
+                    if first {
+                        first = false
+                    } else {
+                        f.write_str("][");
+                    }
+                    f.write_node(p);
+                }
+
                 f.write_str("]");
             }
         }
@@ -629,6 +662,8 @@ impl_display!(HomogenizingFunction);
 pub struct SubscriptPosition<T: AstInfo> {
     pub start: Option<Expr<T>>,
     pub end: Option<Expr<T>>,
+    // i.e. did this subscript include a colon
+    pub explicit_slice: bool,
 }
 
 impl<T: AstInfo> AstDisplay for SubscriptPosition<T> {
@@ -636,9 +671,11 @@ impl<T: AstInfo> AstDisplay for SubscriptPosition<T> {
         if let Some(start) = &self.start {
             f.write_node(start);
         }
-        f.write_str(":");
-        if let Some(end) = &self.end {
-            f.write_node(end);
+        if self.explicit_slice {
+            f.write_str(":");
+            if let Some(end) = &self.end {
+                f.write_node(end);
+            }
         }
     }
 }